Injection SQL/JS (Done !!!)

+++ App/Backend/Modules/News/NewsController.php
@@ -10,9 +10,9 @@ namespace App\Backend\Modules\News;
 
 use \OCFram\BackController;
 use \OCFram\HTTPRequest;
-use \OCFram\Comment;
+use \Entity\Comment;
 use \Entity\News;
-use \OCFRam\FormHandler;
+use \OCFram\FormHandler;
 use \FormBuilder\NewsFormBuilder;
 use \FormBuilder\CommentFormBuilder;
 
@@ -41,8 +41,14 @@ class NewsController extends BackController
      */
     public function executeInsert(HTTPRequest $request)
     {
-        $this->processForm($request);
-        $this->page->addVar('title', 'Ajout d\'une news');
+        if($this->app->user()->isAuthenticated()) {
+            $this->processForm($request);
+            $this->page->addVar('title', 'Ajout d\'une news');
+        }
+        else {
+            $this->app->user()->setFlash('Vous n\'etes pas authentifié pour effectuer cette action..');
+            $this->app->httpResponse()->redirect('.');
+        }
     }
 
     /**
@@ -50,24 +56,40 @@ class NewsController extends BackController
      */
     public function executeUpdate(HTTPRequest $request)
     {
-        $this->processForm($request);
-        $this->page->addVar('title', 'Modification d\'une news');
+        if($this->app->user()->isAuthenticated()) {
+            $this->processForm($request);
+            $this->page->addVar('title', 'Modification d\'une news');
+        }
+        else {
+            $this->app->user()->setFlash('Vous n\'etes pas authentifié pour effectuer cette action..');
+            $this->app->httpResponse()->redirect('.');
+        }
     }
 
     /**
+     * Supprimer une news
      * @param HTTPRequest $request
      */
     public function executeDelete(HTTPRequest $request)
     {
-        $newsId = $request->getData('news_id');
-        var_dump($request->getData('news_id'));
-        die();
-        $this->managers->getManagerOf('News')->delete($newsId);
-        $this->managers->getManagerOf('Comments')->deleteFromNews($newsId);
+        if($request->getExists('news_id')) {
+            $newsId = $request->getData('news_id');
+        }
 
-        $this->app->user()->setFlash('La news a bien été supprimée !');
+        $news = $this->managers->getManagerOf('News')->getNewscUniqueUsingNewsId($newsId);
 
-        $this->app->httpResponse()->redirect('.');
+        if($this->app->user()->isAuthenticated() && ($this->app->user()->rule() == \Entity\Author::RULE_ADMIN | $this->app->user()->sessionUser() == (int) $news->FNC_fk_FAC())) {
+            $this->managers->getManagerOf('News')->delete($newsId);
+            $this->managers->getManagerOf('Comments')->deleteFromNews($newsId);
+
+            $this->app->user()->setFlash('La news a bien été supprimée !');
+
+            $this->app->httpResponse()->redirect('.');
+        }
+        else {
+            $this->app->user()->setFlash('Vous n\'avezpas les droits pour supprimer cette news !');
+            $this->app->httpResponse()->redirect('.');
+        }
     }
 
     /**
@@ -76,24 +98,27 @@ class NewsController extends BackController
      */
     public function processForm(HTTPRequest $request)
     {
-        if($request->method() == 'POST')
+        if ($request->method() == 'POST')
         {
             $news = new News([
                 'FNC_title' => $request->postData('FNC_title'),
+                'FNC_fk_FAC' => $this->app->user()->sessionUser(),
                 'FNC_content' => $request->postData('FNC_content')
             ]);
 
-            if($request->getExists('FNC_id'))
+            if ($request->getExists('id'))
             {
-                $news->setId($request->postData('FNC_id'));
+                $news->setId($request->getData('id'));
             }
         }
-        else{
-            if($request->getExists('FNC_id'))
+        else
+        {
+            if ($request->getExists('id'))
             {
-                $news = $this->managers->getManagerOf('News')->getNewscUniqueUsingNewsId($request->getData('FNC_id'));
+                $news = $this->managers->getManagerOf('News')->getNewscUniqueUsingNewsId($request->getData('id'));
             }
-            else{
+            else
+            {
                 $news = new News;
             }
         }
@@ -105,13 +130,17 @@ class NewsController extends BackController
 
         $formHandler = new \OCFram\FormHandler($form, $this->managers->getManagerOf('News'), $request);
 
-        if($formHandler->process())
-        {
-            $this->app->user()->setFlash($news->isNew() ? 'La news a bien été ajoutée !' : 'La news a bien été modifiée !');
-            $this->app->httpResponse()->redirect('/admin/');
+        if ($news->isNew() | ($this->app->user()->rule() == \Entity\Author::RULE_ADMIN | $this->app->user()->sessionUser() == (int)$news->FNC_fk_FAC())) {
+            if ($formHandler->process()) {
+                $this->app->user()->setFlash($news->isNew() ? 'La news a bien été ajoutée !' : 'La news a bien été modifiée !');
+                $this->app->httpResponse()->redirect('/admin/');
+            }
+            $this->page->addVar('form', $form->createView());
         }
-
-        $this->page->addVar('form', $form->createView());
+        else {
+            $this->app->user()->setFlash('Vous n\'etes pas auteur de cette news.');
+            $this->app->httpResponse()->redirect('.');
+        }
     }
 
     /**
@@ -121,16 +150,20 @@ class NewsController extends BackController
     public function executeUpdateComment(HTTPRequest $request)
     {
         $this->page->addVar('title', 'modification d\'un commentaire');
-        $fcc_id = $request->getData('id');
 
+        if($request->getExists('id')) {
+            $fcc_id = $request->getData('id');
+        }
+
+        $comment = $this->managers->getManagerOf('Comments')->get($fcc_id);
         if($request->method() == 'POST') {
             $comment = new Comment([
                 'FCC_id' => $fcc_id,
+                'FCC_fk_FAC' => $comment->FCC_fk_FAC(),
+                'FCC_fk_FNC' => $comment->FCC_fk_FNC(),
                 'FCC_content' => $request->postData('FCC_content')
             ]);
-        }
-        else {
-            $comment = $this->managers->getManagerOf('Comments')->get($fcc_id);
+            $comment->setId($fcc_id);
         }
 
         $formBuilder = new CommentFormBuilder($comment);
@@ -140,19 +173,35 @@ class NewsController extends BackController
 
         $formHandler = new FormHandler($form, $this->managers->getManagerOf('Comments'), $request);
 
-        if($formHandler->process())
-        {
-            $this->app->user()->setFlash('Le commentaire a bien été modifié');
-            $this->app->httpResponse()->redirect('/admin/');
+        /*var_dump($comment);
+        var_dump($this->app->user()->sessionUser());
+        die();*/
+        if($this->app->user()->sessionUser() == $comment->FCC_fk_FAC()) {
+            if ($formHandler->process()) {
+                $this->app->user()->setFlash('Le commentaire a bien été modifié');
+                $this->app->httpResponse()->redirect('/news-' . $comment->FCC_fk_FNC() . '.html');
+            }
+            $this->page->addVar('form', $form->createView());
         }
-
-        $this->page->addVar('form', $form->createView());
+        else {
+            $this->app->user()->setFlash('Vous n\'avez aucun droit sur ce commentaire.');
+            $this->app->httpResponse()->redirect('.');
+        }
     }
 
+    /**
+     * @param HTTPRequest $request
+     */
     public function executeDeleteComment(HTTPRequest $request)
     {
-        $this->managers->getManagerOf('Comments')->delete($request->getData('id'));
-        $this->app->user()->setFlash('Le commentaire a bien été supprimé !');
-        $this->app->httpResponse()->redirect('.');
+        if($request->getExists('id')) {
+            $fcc_id = $request->getData('id');
+        }
+        $comment = $this->managers->getManagerOf('Comments')->get($fcc_id);                 //Je recupere le comment par son id
+        if($this->app->user()->sessionUser() == \Entity\Author::RULE_ADMIN | $this->app->user()->sessionUser() == (int)$comment->FCC_fk_FAC()) {
+            $this->managers->getManagerOf('Comments')->delete($fcc_id);
+            $this->app->user()->setFlash('Le commentaire a bien été supprimé !');
+            $this->app->httpResponse()->redirect('/news-' . $comment->FCC_fk_FNC() . '.html');
+        }
     }
 }
